requirements.txt
aenum==3.1.15
anyio==4.4.0
asteval==1.0.1
backoff==2.2.1
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
dataclasses-json==0.6.7
deprecation==2.1.0
dill==0.3.8
funcsigs==1.0.2
gs-quant==1.0.104
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
lmfit==1.3.2
marshmallow==3.21.3
more-itertools==10.3.0
msgpack==1.0.8
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
opentracing==2.4.0
packaging==24.1
pandas==2.1.4
patsy==0.5.6
pydash==6.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
scipy==1.14.0
six==1.16.0
sniffio==1.3.1
statsmodels==0.14.2
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uncertainties==3.2.2
urllib3==2.2.2
websockets==12.0
